requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
supabase>=2.0.3
pydantic>=2.4.2,<3.0.0
pydantic-settings>=2.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
typing-extensions>=4.8.0
email-validator>=2.1.0 